# docs/apps/openapi.yaml
openapi: 3.1.0
info:
  title: API de Umbani para las apps
  version: 1.0.0
  description: |
    La app PINTA, nunca calcula: todo importe llega del servidor.
    Reglas completas en docs/apps/API-UMBANI.md. Tipos detallados de la carta
    y del pedido en apps/store/src/lib/types.ts (la mini app web usa esta misma API).
servers:
  - url: https://web-production-3433c.up.railway.app
    description: Producción (NO desarrollar contra ella)
components:
  securitySchemes:
    sesionApp:
      type: http
      scheme: bearer
      description: Token de POST /api/v1/auth/whatsapp/verificar. Dura 30 días.
    sesionTienda:
      type: apiKey
      in: header
      name: x-storefront-token
      description: Token de POST /api/v1/locales/{slug}/sesion. Va SIEMPRE con x-storefront-device, User-Agent y Accept-Language fijos.
  parameters:
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
    dispositivo:
      name: x-storefront-device
      in: header
      required: true
      schema: { type: string, minLength: 8 }
...
paths:
...
  /api/v1/auth/whatsapp/verificar:
    post:
      summary: Canjea el código por la sesión de la app (preguntar cada 2–3 s)
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [codigo], properties: { codigo: { type: string } } }
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema: { type: object, properties: { token: { type: string }, telefono: { type: string } } }
        '202': { description: 'Todavía no llegó el mensaje: { pendiente: true }' }
        '410': { description: El código venció o ya se usó }
  /api/v1/yo:
    get:
      summary: El teléfono de la sesión
      security: [{ sesionApp: [] }]
      responses:
        '200': { description: '{ telefono }' }
        '401': { description: Sesión vencida }
...
  /api/v1/locales/{slug}/sesion:
    post:
      summary: La sesión de tienda para entrar en un local
      security: [{ sesionApp: [] }]
      parameters: [{ $ref: '#/components/parameters/slug' }, { $ref: '#/components/parameters/dispositivo' }]
      responses:
        '201': { description: '{ token } para x-storefront-token' }
        '400': { description: Falta x-storefront-device }
        '404': { description: No existe ese local }
        '409': { description: El local no recibe pedidos }
...